Extract variadic functions
Extract variadic functions to their own file. No behavior change expected.
